# API_Docs_Swagger_Final.yaml
openapi: 3.0.3
info:
  version: '1.0'
  title: Bug Bounty Platform 
  description: |-
    El objetivo de este projecto es crear una API que permita a las distintas empresas crear múltiples programas de Bug Bounty y mediante estos, realizar una gestión de las vulnerabilidades envíadas por los Pentesters.
    Los usuarios en la plataforma son los siguientes:
    - Usuario Hacker: Usuario que únicamente puede envíar vulnerabilidades y ver programas de Bountys existentes. Este es un usuario que puede ser creado públicamente (Mediante los endpoints de Users)
    - Usuario Empresa: Usuario que puede solamente ver y modificar información sobre su propia empresa, crear programas de Bounties y visualizar vulnerabilidades de los programas pertenecientes únicamente de su empresa. Este usuario debe ser creado desde el Backend. Estos usuarios serán los encargados de decidir si la vulnerabilidad es válida o no.
    - Usuario SuperAdmin: Este usuario puede administrar completamente la plataforma.
tags:
  - name: users
    description: Operaciones del usuario
  - name: companies
    description: Operaciones de la empresa
    
paths:
  /v1/users/login:
      post:
        tags:
          - users
        summary: Inicio de sesión en el sistema
        description: 'Inicio de sesión'
        parameters:
          - name: username
            in: query
            description: Nombre de usuario
            required: false
            schema:
              type: string
          - name: password
            in: query
            description: Clave del usuario
            required: false
            schema:
              type: string
        responses:
          '200':
            description: Inicio de sesión exitoso
            headers:
              X-Rate-Limit:
                description: Limitación de intentos por usuarios permitidos por hora
                schema:
                  type: integer
                  format: int32
              X-Expires-After:
                description: Fecha de expiración del token / cookie
                schema:
                  type: string
                  format: date-time
          '400':
            description: Usuario o clave inválida
  /v1/users:
    post:
      tags:
        - users
      summary: Creacion de un usuario en el sistema
      description: Creacion de un usuario en el sistema
      responses:
          '200':
            description: Creación del usuario exitosa
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
  /v1/users/{username}:
    put:
      tags:
        - users
      summary: Actualizar la información del usuario
      description: Actualizar la información de un usuario
      parameters:
        - name: username
          in: path
          description: Nombre del usuario a modificar
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Modificación del usuario exitosa
        '400':
          description: El nombre de usuario proporcionado es invalido
        '401':
          description: Usuario sin privilegios para ejecutar la acción
        '404':
          description: Usuario no encontrado
        '500':
          description: Se ha producido un error en el servidor
    delete:
      tags:
        - users
      summary: Eliminar el usuario
      description: Eliminar un usuario de la plataforma
      parameters:
        - name: username
          in: path
          description: Nombre del usuario a eliminar
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Creación del usuario exitosa
        '400':
          description: El nombre de usuario proporcionado es invalido
        '401':
          description: Usuario sin privilegios para ejecutar la acción
        '404':
          description: Usuario no encontrado
        '500':
          description: Se ha producido un error en el servidor
  /v1/companies:
    get:
      tags:
        - companies
      summary: Obtener una lista con todas las empresas
      description: Obtener una lista con todas las empresas. Solo ejecutable por un SuperUsuario.
      responses:
          '200':
            description: Creación de la empresa exitosa
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
    post:
      tags:
        - companies
      summary: Crear empresa
      description: Permite crear una nueva empresa. Solo ejecutable por un SuperUsuario.
      responses:
          '200':
            description: Creación de la empresa exitosa
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
  /v1/companies/{companyid}:
    get:
        tags:
          - companies
        summary: Obtener informacón de la empresa
        description: Permite obtener información de la empresa.
        parameters:
          - name: companyid
            in: path
            description: Company ID
            required: true
            schema:
              type: string
        responses:
          '200':
            description: Información de la empresa
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
    put:
      tags:
        - companies
      summary: Actualizar información de la empresa
      description: Permite modificar la información de una empresa. Solo ejecutable por un SuperAdmin.
      parameters:
          - name: companyid
            in: path
            description: Company ID
            required: true
            schema:
              type: string
      responses:
          '200':
            description: Modificación de la empresa exitosa
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
    delete:
      tags:
        - companies
      summary: Eliminar la empresa
      description: Permite eliminar la empresa. Solo ejecutable por un SuperAdmin.
      parameters:
          - name: companyid
            in: path
            description: Company ID
            required: true
            schema:
              type: string
      responses:
          '200':
            description: Modificación del programa exitoso
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
  /v1/companies/{companyid}/vulns:
    get:
        tags:
          - companies
        summary: Obtener una lista con todas las vulnerabilidades del programa
        description: Permite ver una lista de todas las vulnerabilidades del programa. Solo ejecutable con un usuario empresa (Con los permisos únicamente sobre los programas que administra) o SuperAdmin.
        parameters:
          - name: companyid
            in: path
            description: Program ID
            required: true
            schema:
              type: string
        responses:
          '200':
            description: Visualización exitosa del listado
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servido
    post:
      tags:
        - companies
      summary: Crear una nueva vulnerabilidad
      description: Permite crear una nueva vulnerabilidad para ser cargada en el programa.
      parameters:
          - name: companyid
            in: path
            description: Program ID
            required: true
            schema:
              type: string
      responses:
          '200':
            description: Creación de la vulnerabilidad exitosa
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
  /v1/companies/{companyid}/vulns/{vulnid}:
    get:
        tags:
          - companies
        summary: Visualizar una vulnerabilidad
        description: Permite visualizar los detalles de una vulnerabilidad puntual. Un usuario \'hacker\' solo puede ver sus vulnerabilidades enviadas, el usuario empresa solo puede visualizar las vulnerabilidades enviadas a los programas que administra.
        parameters:
          - name: companyid
            in: path
            description: Program ID
            required: true
            schema:
              type: string
          - name: vulnid
            in: path
            description: Vuln ID
            required: true
            schema:
              type: string
        responses:
          '200':
            description: Detalles de la vulnerabilidad
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
  /v1/companies/{companyid}/vulns/{vulnid}/approve:
    get:
        tags:
          - companies
        summary: Cambiar el estado de una vulnerabilidad a Aprobada
        description: Permite cambiar el estado de una vulnerabilidad a aprobado. Solo ejecutable con un usuario empresa (Con los permisos únicamente sobre los programas que administra) o SuperAdmin.
        parameters:
          - name: companyid
            in: path
            description: Program ID
            required: true
            schema:
              type: string
          - name: vulnid
            in: path
            description: Vuln ID
            required: true
            schema:
              type: string
        responses:
          '200':
            description: La vulnerabilidad fue aprobada
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor
  /v1/companies/{companyid}/vulns/{vulnid}/reject:
    get:
        tags:
          - companies
        summary: Cambiar el estado de una vulnerabilidad a Rechazada
        description: Permite cambiar el estado de una vulnerabilidad a rechazado. Solo ejecutable con un usuario empresa (Con los permisos únicamente sobre los programas que administra) o SuperAdmin.
        parameters:
          - name: companyid
            in: path
            description: Program ID
            required: true
            schema:
              type: string
          - name: vulnid
            in: path
            description: Vuln ID
            required: true
            schema:
              type: string
        responses:
          '200':
            description: La vulnerabilidad fue rechazada
          '401':
            description: Usuario sin privilegios para ejecutar la acción
          '500':
            description: Se ha producido un error en el servidor